Implemented move history Rewrote Sudoku to be a class instead of struct Reworked logic for adding moves

+++ Simple Sudoku/SudokuCell.swift
@@ -63,7 +63,7 @@ class SudokuCell: UICollectionViewCell {
         }
     }
     
-    func setNumber(to num: Int) {
+    func setNumber(to num: Int8) {
         self.cellLabel.text = num != 0 ? String(num) : ""
     }
